+++ src-docs/src/views/form_layouts/described_form_group.js
@@ -87,8 +87,7 @@ export default class extends Component {
     return (
       <EuiForm>
         <EuiDescribedFormGroup
-          title={<h3>Single text field</h3>}
-          legend="Single text field"
+          title="Single text field"
           description={
             <Fragment>
               A single text field that can be used to display additional text.
@@ -104,17 +103,14 @@ export default class extends Component {
           </EuiFormRow>
         </EuiDescribedFormGroup>
 
-        <EuiDescribedFormGroup
-          title={<h3>No description</h3>}
-          legend="No description">
+        <EuiDescribedFormGroup title="No description">
           <EuiFormRow label="Text field">
             <EuiFieldText name="first" />
           </EuiFormRow>
         </EuiDescribedFormGroup>
 
         <EuiDescribedFormGroup
-          title={<h3>Multiple fields</h3>}
-          legend="Multiple fields"
+          title="Multiple fields"
           description="Here are three form rows. The first form row does not have a title.">
           <EuiFormRow
             hasEmptyLabelSpace
@@ -139,8 +135,7 @@ export default class extends Component {
         </EuiDescribedFormGroup>
 
         <EuiDescribedFormGroup
-          title={<h2>Full width</h2>}
-          legend="Full width"
+          title="Full width"
           titleSize="xxxs"
           description={
             <Fragment>